requirements.txt
actionlib==1.14.0
angles==1.9.13
asgiref==3.3.1
catkin==0.8.10
controller-manager==0.19.6
controller-manager-msgs==0.19.6
defusedxml==0.7.1
diagnostic-updater==1.11.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
dynamic-reconfigure==1.7.3
Faker==4.14.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
laser-geometry==1.6.7
message-filters==1.15.15
moveit-commander==1.1.11
moveit-core==1.1.11
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.11
pr2-power-board==1.1.10
py-trees==0.7.6
python-dateutil==2.8.1
python-qt-binding==0.4.4
pytz==2020.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource-retriever==1.12.7
rosbag==1.15.15
rosclean==1.15.8
rosgraph==1.15.15
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospy==1.15.15
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-controller-manager==0.19.6
rqt-dep==0.4.12
rqt-ez-publisher==0.6.1
rqt-graph==0.4.14
rqt-ground-robot-teleop==1.4.2
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-joint-trajectory-controller==0.21.0
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-multiplot==0.0.12
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-pr2-dashboard==0.4.0
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-py-trees==0.4.1
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.19
sensor-msgs==1.13.1
six==1.15.0
smclib==1.8.6
sqlparse==0.4.4
srdfdom==0.6.3
text-unidecode==1.3
tf==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
topic-tools==1.15.15
unique-id==1.0.6
urdfdom-py==0.4.6
validate-docbr==1.8.2
xacro==1.14.14
